+++ cpp/src/io/csv/csv_reader.cu
@@ -40,6 +40,11 @@
 #include <sys/stat.h>
 #include <sys/mman.h>
 
+#include <memory>
+#include <arrow/status.h>
+#include <arrow/io/interfaces.h>
+#include <arrow/io/file.h>
+
 #include <thrust/scan.h>
 #include <thrust/reduce.h>
 #include <thrust/device_ptr.h>
@@ -224,6 +229,80 @@ __device__ void setBit(gdf_valid_type* address, int bit) {
 	return GDF_SUCCESS;
 }
 
+/**
+ * reads contents of an arrow::io::RandomAccessFile in a char * buffer up to the number of bytes specified in bytes_to_read
+ * for non local filesystems where latency and availability can be an issue it will retry until it has exhausted its the read attemps and empty reads that are allowed
+ */
+gdf_error read_file_into_buffer(std::shared_ptr<arrow::io::RandomAccessFile> file, int64_t bytes_to_read, uint8_t* buffer, int total_read_attempts_allowed, int empty_reads_allowed){
+
+	if (bytes_to_read > 0){
+
+		int64_t total_read;
+		arrow::Status status = file->Read(bytes_to_read,&total_read, buffer);
+
+		if (!status.ok()){
+			return GDF_FILE_ERROR;
+		}
+
+		if (total_read < bytes_to_read){
+			//the following two variables shoudl be explained
+			//Certain file systems can timeout like hdfs or nfs,
+			//so we shoudl introduce the capacity to retry
+			int total_read_attempts = 0;
+			int empty_reads = 0;
+
+			while (total_read < bytes_to_read && total_read_attempts < total_read_attempts_allowed && empty_reads < empty_reads_allowed){
+				int64_t bytes_read;
+				status = file->Read(bytes_to_read-total_read,&bytes_read, buffer + total_read);
+				if (!status.ok()){
+					return GDF_FILE_ERROR;
+				}
+				if (bytes_read == 0){
+					empty_reads++;
+				}
+				total_read += bytes_read;
+			}
+			if (total_read < bytes_to_read){
+				return GDF_FILE_ERROR;
+			} else {
+				return GDF_SUCCESS;
+			}
+		} else {
+			return GDF_SUCCESS;
+		}
+	} else {
+		return GDF_SUCCESS;
+	}
+}
+
+
+/**
+ * @brief read in a CSV file
+ *
+ * Read in a CSV file, extract all fields, and return a GDF (array of gdf_columns) using arrow interface
+ **/
+
+gdf_error read_csv_arrow(csv_read_arg *args, std::shared_ptr<arrow::io::RandomAccessFile> arrow_file_handle)
+{
+ 	void * 		map_data = NULL;
+	int64_t 	num_bytes;
+	arrow_file_handle->GetSize(&num_bytes);
+	map_data = (void *) malloc(num_bytes);
+	gdf_error error = read_file_into_buffer(arrow_file_handle, num_bytes, (uint8_t*) map_data,100,10);
+	checkError(error, "reading from file into system memory");
+
+	args->input_data_form = gdf_csv_input_form::HOST_BUFFER;
+	args->filepath_or_buffer = (const char *)map_data;
+	args->buffer_size = num_bytes;
+	
+	error = read_csv(args);
+	free(map_data);
+
+	//done reading data from map
+	arrow_file_handle->Close();
+
+	return error;
+}
 
 /**---------------------------------------------------------------------------*
  * @brief Read in a CSV file, extract all fields and return 
@@ -1729,5 +1808,4 @@ __device__ int findSetBit(int tid, long num_bits, uint64_t *r_bits, int x) {
 
 	return offset;
 }
-